crypto: Cleanup crypto_x509_is_CA
Going to be doing this for all crypto functions so going to
summarize changes here. Basically the previous way of defining
crypto functions involved 2 extra layers of function pointers
before actually getting the implementation. I don't know why
the entire crypto_util.c is needed and not sure why so many
different wrappers areound the same function are needed. So,
just call the function directly.
Add documentation so we have a consistent API when we add new
crypto library support.

Signed-off-by: Nick Child <nick.child@ibm.com>