crypto: Redefine x509 get len functions
Several fixups to the *get_len functions. Mainly the API is
different so that the return code is the length if positive
and an error code on negative. This aligns with secvarctl
and fixes the following bug which could return success when
the length is 0:
  rc = crypto_x509.get_sig_len (x509, &bit_len);
  if (rc != SV_SUCCESS || bit_len == 0)
    return rc;

And, similar to the previous commit, call the crypto functions
directly instead of through various wrappers. Also, add documentation
on crypto functions.

Signed-off-by: Nick Child <nick.child@ibm.com>