crypto: Introduce CRYPTO_SUCCESS value, redefine error_string
This macro can be used by callers of cypto functions to
determine if the crypto function returns a success or error value.
Some crypto libraries like to return 0 on succes, others prefer 1.

We are trying to pass as many crypto error codes down to the callers
as possible. This increases transparency between the callers of the
crypto functions and the crypto libraries themselves. Also, being
honest, the libraries use more robust and useful return codes anyways.

The user can use crypto_strerror (prvsly error_string) to get more
information. This is used in secvarctl but not libstb-secvar.

Signed-off-by: Nick Child <nick.child@ibm.com>