crypto: Redefine crypto_x509_oid_is_pkcs1_sha256
So this function is the logically the same as crypto_x509_md_is_sha256.
crypto_x509_oid_is_pkcs1_sha256 is not used in libstb-secvar but it is
used in secvarctl. The two functions are defined differently in the
mbedtls implementation so we should make a note to come back and
fix the openssl implementation.

Signed-off-by: Nick Child <nick.child@ibm.com>