crypto: Treat PKCS7s as PKCS7s instead of PKCS7_SIGNED
I understand why PKCS7_SIGNED_free was being used now.

Apparently the pkcs7 structure being used was PKCS7_SIGNED
instead of openssl's PKCS7 strucure. I believe this was because
sbvarsign will generate PKCS7_SIGNED instead of the whole
pkcs7 structure. So by parsing all pkcs7 DER buffers as
PKCS7_SIGNED, compatability with both secvarctl and sbvarsigned
is preserved.

The problem with this is that it is misleading to the user (who
thinks they are dealing with the full structure) and when
other crypto lib support is added, there will be further confusion
as to what a "crypto_pkcs7_t" is.

Instead of casting every pkcs7 buffer into PKCS7_SIGNED, cast them
into PKCS7 structures.

Signed-off-by: Nick Child <nick.child@ibm.com>