crypto: Redefine pkcs7_md_is_sha256
Remove the usual fluff and toss in a possible memleak fix.
Trying to keep track of propogating misleading return codes.
We want to return as much crypto lib return codes as possible.
I think gnutls will return 1 on success so we need to try to
minimize the pain of adding that in.

Signed-off-by: Nick Child <nick.child@ibm.com>