crypto: Redefine md_generate_hash
Previously a wrapper was used so that the returned hash
could be allocated before calling the generate function.

Sounds like a bunch of cruft to me. So I removed it and
now callers pass an unallocated out ptr directly to
generate, if this really grinds gears then we can modify
md_generate_hash to accept an already allocated ptr.

Also, fix bug in test_auth.c that returned success always.

Signed-off-by: Nick Child <nick.child@ibm.com>