pseries: move variable header struct to pseries.h, rename to include header in name
secvarctl needs to parse the 8-byte header that the variables are stored
with, so the struct that defines that header layout should be accessible
via a header include.

In the meantime, also rename:
 - struct signed_variable -> signed_variable_header
 - struct var_timestamp -> var_hdr_timestamp
to indicate that these structures are used specifically for a header,
and not to be confused with the 16-byte timestamp in ESL/AUTH structures.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>